Refactor SSE emitter handling
