+++ back/src/main/java/co/com/arena/real/application/service/SseService.java
@@ -4,75 +4,19 @@ import co.com.arena.real.infrastructure.dto.rs.TransaccionResponse;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
 import java.io.IOException;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 
 @Service
 @RequiredArgsConstructor
-public class SseService {
+public class SseService extends AbstractSseEmitterService {
 
     private static final Logger log = LoggerFactory.getLogger(SseService.class);
 
-    private static class EmitterWrapper {
-        final SseEmitter emitter;
-        volatile long lastAccess;
-
-        EmitterWrapper(SseEmitter emitter) {
-            this.emitter = emitter;
-            this.lastAccess = System.currentTimeMillis();
-        }
-    }
-
-    private final Map<String, EmitterWrapper> emitters = new ConcurrentHashMap<>();
-
     public SseEmitter subscribe(String jugadorId) {
-        String lock = ("lock_" + jugadorId).intern();
-        synchronized (lock) {
-            EmitterWrapper existing = emitters.remove(jugadorId);
-            if (existing != null) {
-                existing.emitter.complete();
-            }
-
-            SseEmitter emitter = new SseEmitter(Long.MAX_VALUE);
-            EmitterWrapper wrapper = new EmitterWrapper(emitter);
-
-            emitter.onCompletion(() -> removeEmitter(jugadorId));
-            emitter.onTimeout(() -> removeEmitter(jugadorId));
-            emitter.onError(e -> removeEmitter(jugadorId));
-
-            emitters.put(jugadorId, wrapper);
-            log.info("Nueva conexión SSE para jugador: {}", jugadorId);
-            return emitter;
-        }
-    }
-
-    @Scheduled(fixedRate = 15000)
-    public void sendHeartbeats() {
-        emitters.forEach((id, wrapper) -> {
-            try {
-                wrapper.emitter.send(SseEmitter.event().comment("heartbeat"));
-                wrapper.lastAccess = System.currentTimeMillis();
-            } catch (IOException e) {
-                removeEmitter(id);
-            }
-        });
-    }
-
-    @Scheduled(fixedRate = 60000)
-    public void limpiarEmitters() {
-        long now = System.currentTimeMillis();
-        long ttl = 5 * 60 * 1000L;
-        emitters.forEach((id, wrapper) -> {
-            if (now - wrapper.lastAccess > ttl) {
-                removeEmitter(id);
-                wrapper.emitter.complete();
-            }
-        });
+        return super.subscribe(jugadorId);
     }
 
     public void notificarTransaccionAprobada(TransaccionResponse dto) {
@@ -111,10 +55,5 @@ public class SseService {
             removeEmitter(jugadorId);
             wrapper.emitter.completeWithError(e);
         }
-    }
-
-    private void removeEmitter(String jugadorId) {
-        emitters.remove(jugadorId);
-        log.info("Desconectado SSE jugador: {}", jugadorId);
     }
 }
